#493 added full thermal model to stucture
