+++ pybamm/models/submodels/thermal/isothermal/isothermal_1D_current_collector.py
