+++ pybamm/models/submodels/thermal/isothermal/isothermal_2D_current_collector.py
