+++ pybamm/models/submodels/thermal/isothermal/isothermal_no_current_collector.py
